RAG/requirements.txt
langchain==0.3.19
langchain-community==0.3.18
langchain-openai==0.3.7
faiss-cpu==1.10.0
openai==1.65.2
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
pypdf==5.3.0
pydantic==2.10.6
redis==5.0.3
python-multipart==0.0.20